Create solution.py
Since O(1) space will be used, for 32-bit integers, there will be enough space to store two values. The integers are all <= 1000, so that is 2^10 or 10 bits. the x and y values each can be fully represented with 10 bits.

First, left-shift by 10 bits for the x values from the first half, then logic OR with the complement to add the y values from the latter half.

Then, initialize j pointer to start from the end. Afterwards, in the first half, logic AND with the sequence of 10 1's in order to get the y value from the last 10 bits, then right shift to get back the x value. These are then stored from the end in their respective spaces, then decremented by 2 spaces.

This saves space by using O(1) space, and has O(n) time complexity.