Create solution.py
time: O(n*k + m) where n is the max length of each word, k is the number of words in the list, m is the length of chars str

space: O(n + m)

first, create a counter of chars for each character. Then, iterate through each word in words, and create a defaultdict with a boolean `good` variable set to True initially.

for each character in the word, iterate, and if not good (character not in chars_count or count is greater in the current word), then break out.

if still good after running through each character of a particular word, add the length to the output accumulator.