Create solution_iter.py
O(n) time and O(h) space complexity iterative solution - h = log(n) for BST

while stack or pointer is not null, iterate. If curr pointer is not null, append the value to the output, and append the right child of the curr node to the stack. Then traverse (update current pointer) to the left child.

If the curr pointer is null (else statement), then pop the stack and assign to the current pointer