Create solution_bfs.py
BFS solution of O(n) time and O(h = log n) space with queue

initialize queue with depth == 1 if root is not empty. Then, while the queue is not empty, iterate through the length of the queue. Pop and make the root the new node from FIFO queue.

Then, if this node has no left or right children, return the depth (this is minimum depth). Otherwise, append the left child and/or right child to the queue.

This will always terminate at a leaf node.