Create solution_dfs_iter.py
Iterative DFS solution with O(n) time and O(h = log(n)) space (O(n) worst-case).

The stack is initialized with the root and level of 1. The output is initialized to a very large number.

While the stack is not empty, pop the node and depth, and if the node has no children, update the min depth output to the minimum of the current value and the current depth.

Append the left and right children to the stack with + 1 depth.