Create solution.py
DFS O(n) time and O(1) space solution

first declare output to be a list - this is for python object being able to be updated in inner function.

Update height of null root to be -1, then perform DFS recursively on left and right subtrees. Add 2 + height of each left and right subtree, and update max diameter with current max and current diameter.

For dfs, return 1 + the max height of left and right subtrees (to keep track of height for diameter calculation in recursion).