Create solution.py
recursive O(n) solution with O(n) space for nodes - checks if root is not available and returns False. Otherwise, check if both root.left and root.right exists, and check if root.val == target sum. Otherwise, return the hasPathSum of the right subtree with targtSum - root.val or hasPathSum of the same for the left subtree.