Create solution.py
preorder traversal - O(n) time and space. return if null root, then append opening parenthesis and root value. However, if the left node is null (but not the right node), add open/closing parentheses for distinction from when only left node (but not right node) exists.

then preorder through left and right subtrees, respectively.

append the closing parentheses here.

finally, join the string of the result array but include 2nd up to (but not including) the last value (since '(' and ')' are added around the whole string, which is not what is needed)