Create solution_dp.py
DP approach

for t in range(1, 3) (which is the at most two transactions limit), set position and profit initially. Then, going through the prices array, calculate the max position from the dp array at index corresponding to the prices at that index, and max profit. Include this max profit in the dp array with respective index.

this is O(n) time and O(n) space