Create solution.py
O(log n) operation for length of n and O(1) space. Zeroing out the most significant bit takes 2^k operations, the next significant bit 2^(k-1), ... all the way to 1. The sum of this is 1 + 2 + 2^2 + ... + 2^k = 2^(k+1) - 1

Then, for all of the non-power-of-2 digits, it can be done by reverting to the previous step at the 2^k power. This can be gotten by XORing the 2^k power number with the number n itself.