Update solution.py
O(n^2) time and O(1) space (in-place)

since the values are between 0 to 255 for each cell, store the average value beyond the least significant 8 bits. This requires modding by 256 in order to get the front most significant bits, as well as left/right shifting appropriately with XOR for the average value stored at each cell value.