Create solution.py
O(n * l * l) time complexity for n being the number of words, where l is the average length of each word. This is due to the loop of dfs for each word, and within each dfs run, it goes through the average length l of each word, and when forming the predecessor, this also takes l length of time due to string slicing. This includes the recursive O(n) calls. The space complexity is O(n) due to the dp[] caching and the word_index[] dictionary.

First, the words are sorted by length in decreasing order (O(n log n) and then a word_index mapping each word to the index it is at within the words list is initialized.

A dp array for caching the backtracking solutions (to lessen recursive calls) is also created.

Within the dfs() call, it first checks if it is in the dp map. Initialize the res to 1, then for each word up to the words in the index, get the predecessors by deleting a random character. If the predecessor is in the word_index, then update the max result.

The O(n) loop of dfs (which gives O(n*l^2) complexity overall) is done, then the max value of the memoized dp.values() is returned.