Create solution_dfs.py
O(n) time visiting all rooms and O(n) space due to recursive callstack.

Have a hashset of all the visited rooms. Add the first room (Room 0) since it is unlocked.

Then, for each key in the room, if the key to the room is already seen, then continue. Otherwise, add this room key to the hashset of seen rooms then perform dfs on that key.

To start, dfs on room 0, and return if the size of the seen rooms is equal to the number of rooms.