Update solution.py
O(n log n + m log m) solution still, where n is the len(flowers), and m is the len(people). O(n + m) space. Now, use pointer j for the flowers array and add to the end heap only. This uses only 1 heap and does not require a count variable, as the length of the heap shows the number of flowers person sees.