Create solution.py
Backtracking solution to go through the matchstick configurations. first, if average length is not divisible by 4, return False.

Create the sides array representing each side of a square. Sort the matchsticks in descending order to do a greedy approach to matchstick configuration (lay longest matchstick first).

Backtrack through each matchsticks - if index reaches end, it means a correct configuration was found - return True.

For each side j in sides, if the current side length + current matchstick length satisfies <= avg_lengths, add it. Then, backtrack function recursively to next index of matchsticks.

Then, take out current index matchstick.

Return False if this does not return True.

This is O(4^n), since 4 branches are created at backtracking. The call stack is recursively of length n (which is the height of this quad tree), so it is O(n) space of recursive calls.