Create solution.py
O(n log n) time complexity due to sorting of values. O(n) space from dictionary and set and sorted values.

Check the counter of each character, and if the set of keys match each other, and the sorted values of each dictionary counts (which means there is a mapping between number of character counts in word1 to word2), the operations can transform from one to the other.