Create solution.py
O(n) time and O(1) space. Create a defaultdict to store the number of ones at each index in range 32 for a 32-bit integer (which should be sufficient for this problem).

Go through the numbers and find the right-shifted i-th number of ones for each number. Initialize a result value of 0.

Then, going through the 32 bits, find the k-th bit for k right shifted to i, and if it is a 1, and the number of ones in the i-th bit is even, it flips so add to result. On the otherhand, if the k-th bit is 0 and the number of ones in that place for the numbers array is odd, then also need to flip so add to the result.