Create solution_kmp.py
O(m + n) time and O(n) space algorithm - KMP algorithm

m = len(haystack)
n = len(needle)

First, calculate longest prefix-suffix array of length n

This is done by using two pointers, one to previous LPS, and one to increment through the needle

After LPS array is initialized, do a similar algorithm through haystack with 2 pointers initialized to 0, and go through the haystack with i and lps with j. If j is equal to length of needle, return right away as this means the needle was found. The index will be the pointer i position - the length of the needle/LPS array.

Otherwise, if not found, return -1