Create solution_treesort_recur.py
Recursive O(n log n) solution with O(n) extra space for tree and output array.

insert() inserts to left or right child recursively for the tree. This is iterated across the given nums list after setting root to the first element.

Next, an inorder traversal through this binary search tree that was built is output into the output array.