Create solution.py
x_vec, y_vec are pointing in 0, 1 for north. Then, x, y are at point 0, 0

For each d in instructions, either add if 'G', swap directions and negate x for 'L', and swap directions and negate y for 'R'. If the end position is (0, 0) or direction has been swapped, then the robot is bounded inside a circle.

O(n) time and O(1) space