Create solution.py
O(nm) time solution where n = number of rows, m = max(number of elements in a given row). O(n) space taken for dictionary of gaps.

First, iterate for each row in the wall and within each row, count the gaps using a prefix sum algorithm that increments the value each iteration. While this is less than the length of a row, add to the count_gaps dictionary. 

Return the length of a row minus the max count (which shows the minimum number of bricks cross), or if the dictionary is empty, return n (as all rows need to be crossed).