Create solution.py
O(26n) = O(n) time solution with O(26) and O(n) space for the right and left palindrome counters

The left palindrome is a set of characters on the left side of the middle pointer (the variable i through which the string is iterated), and the right palindrome is a dictionary mapping the counts of characters to the right of mid.

By using a constant range = 26 inner loop, the characters for the result matching on both the left and right side are added to the result with the character (in order to signify the mid vs. left/right characters), and the length of this result is returned.