Create solution.py
Brute force O((2^n)^2) = O(4^n) solution. 2^n possible strings (include/not include) for the bitmask of the palindrome lengths in the sequence, and ^2 of this value for the nested for loop.

O(2^n) space for the dictionary of palindromes with their respective lengths. The dictionary only contains values if it is a palindrome as the key (the mask of the corresponding palindrome), and the value as the length of said palindrome.

In the mask calculation, the subsequence is added for index i instead of index (n-i-1) due to symmetry.