Create solution.py
O(n) time solution with O(n) space for two prefix arrays. First, create prefix arrays of each of the 2 rows. Then, iterate through. 

The top row defines where robot 1 "turns" to the 2nd row. Thus, calculate from last 1st row element - the i-th element.

For the bottom row, choose the (i-1)-th element and set to 0 for first element of 2nd row.

Now, the 2nd robot to maximize the values of top and bottom rows at each index of turn.  Now, find the minimum of the current result and the value obtained by the second robot (based on the fact that the first robot will try to minimize the score for the second robot)