Create solution.py
O(n) time (possibly going through queue of all dominoes) and O(n) space (due to queue) solution.

First, convert the string into a list of characters.

Then, enumerate and append all elements that are not "." to the queue with its corresponding index. Then, while the queue is not empty, pop and if it is the left (and greater than index 0 while the element to the left (i-1 index) is a ".", append to the queue (turns into an "L") and change the value of that index to "L".

For the right ("R"), if the element immediately to the right (i+1 index) is less than the length of the list and is a ".", as well as the one to its right (i+2 index) is less than the length of the list and is an "L", just pop from the list (since the middle one won't budge). Otherwise, append to the list with (i+1)-th index to "R".

Return the joined string of the dominoes in the list