Create solution.py
O(n) time and space complexity solution.

Use a sliding window of size 10 and only iterate through length - 9 index (only need left pointer). Create HashSets of seen and result. If the 10-string character is in the seen hashset, then add to the result set. Add to seen set each time regardless (if it is a duplicate, it won't matter since it is a set).

Return result as list.

Optimization to consider: since the characters are only in the possibility space of {A, G, C, T}, it only requires 2-bits to encode (as opposed to 8-bit string characters). Further, since the string lengths that are being considered is 10, it is a total of 10 x 2-bits = 20-bits for each string representation. Thus, each 10-letter string can be stored by a 32-bit integer and bit manipulation can achieve this. However, time complexity will still be the same.