Create solution.py
O(1) insert, remove, and getRandom operations.

Use both a hashmap (whose keys are the values, and the values are the indices to the list index which contains those values). This allows for efficient getRandom, as well as the remove function.

The insert() is standard, whereas the remove() method has additional steps in getting the last element of the list which contains the value (by looking up in the hashmap), then setting the idx of the value with the last element of the list while removing (popping) that element from the list (at the end). Then, the dictionary (hashmap) value for last_val is updated to the index, and then the value to remove is deleted. This also works if the hashmap/list only contains one element.

The getRandom() function is just using the built-in random.choice which randomly chooses from an iterable.