Create solution.py
O(n) time solution iterating using sliding window, and O(nk) space for the n*k possible binary codes.

Instead of enumerating all 2^k possible binary codes, go through the string once with a window size k and add to a set. If the set size at the end of the loop is equal to 2^k, then done. Otherwise, false.

This is predicated on the fact that the string is a binary string (or formatted this way).