Create solution.py
O(n) time and O(1) space complexity solution.

Have a res accumulator variable and zero_ctr for counting current number of zeroes. 

Iterate through the list of nums, and if the number is not 0, add to the result the value n * (n+1)/2 where n is the current zero counter. This will not be affected if current zero counter is 0 since the  value will evaluate to 0. Afterwards, set zero_ctr to 0. Continue the loop to skip to next iteration.

If the number == 0, then zero_ctr is incremented by 1. If the 0 is the last element of the list of numbers, then evaluated the sum formula and add to the result.