Create solution.py
O(n) time complexity and O(26) = O(1) space solution for the HashSet (since there are only 26 possible lowercase characters, the HashSet will have at maximum 26 characters).

First, initialize an empty set allocated for characters, along with a counter at 1 (it will skip the last character string partition, so need to start from 1). Then, during the iteration char by char, if the character is found already in the set, then add to the counter and reset the set to an empty one. Throughout each iteration, add the character to the set.

Return the counter result. This approach is greedy and will have a minimum partition.