Create solution_xor.py
O(n^2) time complexity solution (iterating over all possible bits) and O(1) space complexity. This gets time limit exceeded error on larger test cases.

One observation is that assigning 'Y' to '1' and 'N' to '0' as a bitstring, and the possible solution of XOR-ing with a bit representation of what time the store closes and summing all the 1's yields the minimum penalty. E.g., if the store is open until hour 0, it would be 0000, if it is open till hour 1, it would be 1000, 2, its bitstring representation would be 1100.

For example, 

for the first test case, which is 'YYNY', this is represented as '1101'.

For each of the 4 cases at which to close shop,

close @ 0th hr:
1101 XOR 0000 = 1101 -> 1 + 1 + 0 + 1 = 3

close @ 1st hr:
1101 XOR 1000 = 0101 -> 0 + 1 + 0 + 1 = 2

close @ 2nd hr:
1101 XOR 1100 =0010 -> 0 + 0 + 0 + 1 = 1

close @ 3rd hr:
1101 XOR 1110 = 0011 -> 0 + 0 + 1 + 1 = 2

close @ 4th hr:
1101 XOR 1111 = 0010 -> 0 + 0 + 1 + 0 = 1

now, the earliest of these times is the 2nd hour, so this value is chosen.

However, using a simple doubly-nested for loop times out.  The range of the outer and inner loops are dependent on the size of the customers list, so perhaps modifying the bitstring into a bit and doing bit manipulation may be the way to decrease this to O(n) time.