Create solution_xor_prefix_sum.py
O(n) time and O(n) space complexity solution using XOR. Use an array of the penalties incurred when not opening the shop at that hr vs opening the shop at that particular hour i.

The values 'Y' are represented as '1', and 'N' as '0'. Then, an XOR operation in effect (see the solution_xor.py) with the resulting number of 1 bits summed, shows the penalty value.

First, the array size 2 * n is created in order to iterate through the array. Essentially, an AND operation with 1111 is done for the case where the shop is not open (case 0), and the AND operation is done with 0000 for the case where the shop is open (case 1), i.e.,

YYNY -> 1101

case 0 (shop not open):
1101 AND 1111 = 1101

case 1 (shop open):
1101 AND 0000 = 0010

This is represented in an array as:
    0 1 2 3 4
   |----------
0 | 0 1 1 0 1
1 | 0 0 0 1 0

The 0th day is initialized to 0.

Now, it can be a DP array to traverse and calculate the costs, but this would lead to TLE (time limit exceeded) due to a O(n^2) traversal of the array for each case. Thus, it is better to use a similar method as that used to solve the Grid Game Leetcode question, which is to create two rows of prefix sum arrays, as such:

    0 1 2 3 4
   |----------
0 | 0 1 2 2 3
1 | 0 0 0 1 1

Each row is a prefix sum.

Now, by noticing that for each row, the number of hours the shop is open until means the prefix sum of that row and the rest of the 0 row is calculated for each index. This formula is given as 

penalty = prefix_sums[0][n] - prefix_sums[0][i] + prefix_sums[1][i]

Then, the penalty is compared to the current penalty value and then the earliest time for the minimum penalty is updated (only when penalty < min_penalty, since for duplicates, the earlier index is required for the solution).

Optimizations can be made, since the 0 and 1 DP rows are inverses of each other for each bit (1 in one of the index is a 0 in the other, and vice versa).