Create solution.py
O(26^2 n) time solution and O(26) space. The space consists of a hashmap with hashsets as values and characters as keys.

The time complexity is 26^2 since there are only 26 letters in the English alphabet and this will be a double for loop of up to range(26), with n being the size of the ideas (assuming the suffix is stored across all 26 keys, the factor of comparing each value to the other set of key's values is O(cn) with constant c.

After adding each character as a key and each suffix as a set for the value in the dictionary, if the characters match up, continue.

Otherwise, create the suffix intersection counter and subtract from the distinct suffices of the compared characters 1 and 2. Add the multiplication of this value to the result (this double for loop will pass through the character in reverse order ((c1, c2) and (c2, c1), so multiplying by 2 doesn't need to be done.)

Return the result