Create solution.py
O(log n) time and O(1) space solution.

Go through the tree level by level starting at the root as current node. If both p and q values are greater than the current node's, set the current node to its right child. If both p and q values are less than the current node's, set to its left child. Otherwise, it means one is less and the other is greater, or one of the values is equal, in which case the current node is the LCA.