Create solution.py
O(n) time and O(1) space solution (in-place ops).

First, set left and right pointers to 0, then while the right pointer is in-bound, initialize count to 1. Then, while r + 1 is less than the length of the input array nums,  and nums[r] and nums[r+1] are the same, increment both the right pointer and the count.  swap left and right pointer positions on the array, then output the results that was accumulated.