Create solution.py
O(n^2) time complexity solution from iterating right pointer respect to the left pointer (the O(n log n) sorting is dominated by the O(n^2) factor). O(1) space.

First, sort the array (this keeps the same number of subsequence max - min difference). Then, while the sum of the numbers at the left and right pointer are greater than the target and the pointers don't cross, decrement the right pointer. If the pointers are not crossed, add 2^(r - l) (the number of possibilities of the sums) to the result, then mod by the result.