Create solution.py
O(n log n) solution from sorting, linear through the sliding window. O(1) space.

Set left and right pointers to 0, and the result and current sum to 0.

Calculate current running sum by adding in the right pointer element. Then, while the right element multiplied by the length of the current sliding window (of size r - l + 1) is greater than the current sum plus k (which sees if the largest, right pointer value of size window_length = r - l + 1 can be achieved with the current sum and using the k operations), subtract the left pointer value from the current sum and increment the left pointer.

Set the result to the maximum of the result and current window size after the loop is finished, and increment the right pointer.

Finally, return the result.