Create solution.py
O(n) time and O(n) space complexity solution.

First, initialize a count_dict defaultdict to count the number of each occurrence. Initialize the left pointer, and the result value and current total.

Iterating the right pointer window for the fruits, count the right-most fruit and increment to the dictionary with the total being incremented as well.

While the count_dict becomes bigger than 2, decrement the fruit at  the counter on the left pointer and decrement the total. Increment the left pointer accordingly.

If the count_dict becomes 0, then pop that fruit from the dictionary.

At each iteration, set the max of the current running total and result to the result, and return this value at the end.