Create solution.py
O(n) time and O(1) space solution.

First, initialize the vowel set of 5 characters, the curr_sum (running sum), result, and the left pointer. Then, initialize the k-size sum, and set it to result.

For the right pointer, iterate from k to the end, and check if it is in the vowel set and increment the running sum, while subtracting the left pointer. Then, update result and increment the left pointer.

Return the result.