Create solution.py
O(n) time complexity and O(n) space complexity solution (from the two alternating strings).

First, extend the string by itself since removing a character at the beginning and putting it at the end is the same as iterating through the string each time.  Then, generate the 2 alternating strings '010101...' and '101010...' of the same length (2 * n), where n is the length of the string.

Initialize the result value to 2*n (which is the maximum possible operations for flipping necessary) and the difference in the values for the two alternating string types.

Then, use a sliding window with left pointer at 0 and iterate the right pointer. If the position at the right pointer doesn't match with the alternating string type 1 or 2, then increment the difference 1 or 2, respectively. If the window size (r - l + 1) becomes greater than n (meaning we need to subtract the left pointer, then if and only if the left pointer character is different from the alternating string 1 and 2, is the difference 1 and 2 decremented. Increment the left pointer.

Now, if the window size is exactly n, then update the result as the minimum of the current result, diff1, and diff2.