Create solution.py
O(n) time complexity solution with O(1) space for sliding window. First, set the result to the length of the nums array + 1 and the current running sum to 0. Set the left pointer to 0.

Now, iterating through the right pointer, add the current right pointer element to the running sum.

While the running sum is greater than or equal to the target, update the result with the minimum of the window length and the result, and subtract the left pointer element and increment the left pointer.

This can be done since all values are positive.

Return 0 if the result has not been updated, otherwise the result.