Create solution_bs.py
O(log(n) + k) time and O(1) space solution.

First, the left and right pointers are set for binary search, and the value and index of the left-most in the array is recorded.

Then, find the midpoint of the search interval is found, and the current difference (with respect to the midpoint vs. x) and the result's difference (which is the current value - x) is compared. The selection criteria to set the new value and the corresponding index is based on the rules outlined for "nearest": if the absolute difference is less, or given the absolute differences are the same, the index is lower.

The update rules for the midpoint are followed through, with a break when the value equal to x is found.

Next, the left and right pointers are set to the index value for the midpoint found above. Then, iterating through k-1, if the left pointer is 0, then the right pointer is incremented, and otherwise if the right pointer is len(arr) - 1 or the difference beteween x and the value to the left of the left pointer (l-1) is less than the value to the right of the right pointer (r+1), the left pointer is decremented. Otherwise, the right pointer is incremented.

Then, the array slice from l to r + 1 (non-inclusive) is returned.