Create solution.py
O(log(n - k) + k) solution, where n is the size of the array. No extra space, so O(1) space.

Set the left and right pointers to 0 and len(arr) - k. Then, while the pointers don't cross, find the midpoint. If the element to the right of the k-size window (m+k) has a lower difference than the left side of the current pointer at m, then set the left pointer to m+1. Otherwise, set the right pointer to m.

When the loop finishes, the pointers will be set such that the minimum difference respecting the lower index values will be returned.