Create solution.py
Find a sliding window which maximizes the size of the window with target sum equal to the sum of all the numbers minus x. This is an O(n) time solution with O(1) space.

First, initialize the target, left pointer, and the running sum. Set the max window size to -1.

Iterate through the right pointer, and increment the running sum by the value at the right pointer. Now, while the pointers don't cross and the running sum is greater than the target, subtract the value of the left pointer and increment the left pointer.

If the running sum equalizes to the target, update the max window.

Return -1 if the max window size hasn't changed, otherwise the inverse of the max window (length of the array - the size of the max window)