Create solution.py
O(n) time and O(n) space complexity solution using monotonically decreasing queue (deque).

First, initialize array size of (len(nums) - k + 1) for the result. Then, set left and right pointers and initialize the deque.

While the right pointer is in-bounds, while the queue (deque) is not empty and the rightmost element value of nums is less than the right pointer value of nums, pop from the queue.

Append the right pointer index to the queue, and then check if the left pointer is greater than the leftmost index of the queue, and pop from the left of the queue.

If the index of r is greater than k, set the left pointer index of the result array to the nums of the queue. Then, increment the left pointer.

Within the loop, increment the right pointer.

Return the result