Create solution.py
O(n) time and space solution using a stack.

For the asteroids, while the stack is not empty and the asteroid in question is negative (going left) and the top of the stack is positive (asteroid going right), add the sum of the asteroids. If the difference is negative, the leftward asteroid is greater, so pop from the stack. If the difference is positive, then 0-out the asteroid that was going left. Otherwise (both asteroids are equal), then zero-out the asteroid and pop from the stack.

Then, if and only if the asteroid is not 0, push to the stack.

Return the stack after the loop