Create solution.py
Use a monotonic decreasing stack. Complexity of the next() method is O(n) time and O(n) space for the stack. 

Set default span to 1, and while the stack is not empty and the top of the stack has a less than or equal to price, add the span and pop from the stack.

Finally, append the new price and span and return the span