Create solution.py
O(n) time and space solution using stack.

For each character in the string, if the stack is not empty and the character matches the character at the top of the stack, increment by 1 the count. Otherwise, append 1 for the count for the character to the stack.

If the count equals k, then pop from the stack. 

Now, while the stack is not empty, get the top of the stack and append the result string by the character multiplied by its count, then pop from the stack. Return the reversed string