Create solution.py
O(n) time and space complexity solution using a stack, where n is the number of characters in the path string.

First, a string is initialized for the current directory.

Then, for each character in the path string with a '/' appended, the character is checked to see if it is a backslash. If so, then the current string is checked to see if it is the double-dot. If the stack is not empty, then the stack is popped (removing directory).

If the current string is not empty and not a single dot, then the current string is appended to the stack.

The current string is then set to an empty string again.

Otherwise, the characters are appended to the current string.

Return the backslash-joined string with a backslash at the very beginning.