Create solution.py
O(n) time and space complexity solution using a stack.

For the length of the string, if the character is not a closing square bracket, append to the stack. If the character is a closing square bracket, initialize a substring variable as an empty string and while the top of the stack is not an opening bracket put the popped value in front of the substring. There is no need to check if the stack is not empty since it is expected that brackets are well-formed and balanced. Pop the last value (which will be the opening square bracket).

Next, set a variable k and while the stack is not empty and the top of the stack is a digit, put the popped value from the stack and put in front of k.

Now, append the substring k-number of times.

Afterwards, while the stack is not empty, add the popped value from the stack to the front of the result string and return this value.