Create solution.py
O(n) time and space complexity solution. Use a monotonically increasing stack and greedily get rid of larger integers when iterating through the string.

For each character in the string, while k is not 0 and the stack's top element is greater than the current digit, decrement k and pop from the stack. Append to the stack the current digit character.

Afterwards, cut the stack to the length of the stack minus k (to get rid of any trailing values, then join the stack. Return the string of the integer (to get rid of leading zeroes) of the resulting string if it is not empty, otherwise return a 0 for the empty string.