Create solution.py
O(n) time and space complexity solution using a monotonically decreasing stack. The stack stores the number and the minimum to the left of the value in question. 

For the numbers in the array, while the stack is not empty and the stack's top value is less than or equal to the current number in the array, pop from the array (to maintain the monotonic stack).

If the stack is not empty and the number is greater than the current minimum (the number to left of the top of the stack, then return True since the 132 pattern has been found.

Append the number and current minimum to the stack, and update the current minimum.

If the loop exits, return false.