Create solution.py
O(n) time complexity and space complexity solution. The strategy is to keep two dictionaries with a dictionary which contains the values added and their respective counts, as well as a dictionary of counts as keys with stacks for their respective counts. Keep a max count variable as well.

Then, when pushing to the stack, add the value's count and increment. If the new count is greater than the max count, adjust the max count and stacks with a new stack for the new max count.

For popping, pop the value from the stacks dictionary of the max count, and decrement from the count dictionary. Then, if the stack becomes empty, decrement the max count and return the result.