Create solution.py
O(log n) time binary search solution.

Set left and right pointers to 1 and n. Set result to 0 initially.

While the search continues, calculate the midpoint, and use the triangular numbers sum for the number of coins for the midpoint. If the coins is greater than n, then set the right pointer to m - 1; otherwise, set the left pointer to m + 1 and update the result to the maximum of itself and m.

Return the result at the end of the loop