Create solution.py
O(n log M) time solution with O(1) space using binary search, where n is the number of weights and M is the maximum possible value of the binary search (the total sum of the weights).

First, initialize the minimum possible value as the maximum of the weights, and the sum of the weights as the maximum possible value (since a ship with capacity equal to sum of the weights can ship in 1 day).

Then, define a helper function to calculate, given a capacity, if the cargo can be shipped within the number of days. Set the number of ships to 1, and the capacity to be the value passed into the function. Then, for each weight, if the current capacity is less than the weight, increment the number of ships needed and reset the capacity. This check prevents the current capacity from going negative.

Decrease the current capacity by the weight.

Return whether the number of ships is less than or equal to the number of days.

While the pointers don't cross, calculate the capacity (midpoint of binary search), and do the check using the aforementioned function. Then, set the result to the minimum of current result and the capacity, and move the right pointer; otherwise, move the left pointer.

Return the result