Create solution.py
O(log(n)) time and O(1) space solution using modified binary search. 

Set left and right pointers, then while the pointers don't cross, calculate the midpoint. If the midpoint is greater than 0 (not end of left side) and the midpoint is less than the element to its left, search the left side of midpoint (shift right pointer to left of midpoint). If the midpoint is less than the rightmost point and it is less than the value to the right of it, search the right side of midpoint (shift left pointer to the right of midpoint. Otherwise, return the midpoint.