Create solution.py
Use modified binary search on the possible values in the interval (0 to 10^9). This yields a time complexity of O(n log m), where m == 10^9 (the possible values on which binary search will be performed), and n is the number of elements in the array.

A helper function is defined for iterating through the threshold and seeing if the absolute difference meets the threshold, and the count and i pointer is advanced appropriately by 2 if so. Otherwise, the i pointer is incremented by 1 only. If the count and p match up, this function returns true, false otherwise.

Now, the binary search algorithm is run to select from the range of possible values the middle threshold. If the range is valid, then the result is updated with the current middle pointer and then the right pointer is updated. Otherwise, the left pointer is updated. Then the result is returned.