Create solution.py
O(n log k) solution where n is the length of the string s, and k is the length of the removable array.

First, defined a two pointer is_subsequence method to check if subseq is a subsequence of s.

Then, initialize result to 0 and pointers for binary search in the removable array.

While the binary search condition is valid, get the midpoint, then set the removed set to be all removable characters up to (and including) the midpoint. If it is a subsequence, update the result and move left pointer to right of mid. Otherwise, move right pointer to left of mid.

Return the result