Create solution.py
O(n log n + n*w + 3*m) = O(n log n + n*w + m) solution, where n log n is for sorting the products list (n is the size of this list), w is the average word size for iterating through each word of the products list, and m is the number of characters in the searchWord (and 3 is the number of maximum possible times to go through for the top 3 search results). The memory complexity is O(m * w) where m is the search word length and w is the length of the average word).

First, sort the products list and set left and right pointers for binary search. For the length of the search word, check if the left and right side prefixes and length are equivalent, and increment/decrement the left/right pointers, respectively.

Append an empty list and adjust the remaining characters by the width of the window (the size of left pointer -> right pointer sublist).

Iterate for the remaining characters and append to the last added empty list of the result list of lists the words.